Created getAllRealStops.py
It looks into the entries in the database, and based on the Route number, it searches for all stops of that route. It looks to the GPS location of the bus at that time and then searches the closest stop to that location. This is placed as the "real stop". Also, the L2 distance is added as another column